Added devise User model and configured webpacker for verbose output
'username' was added to User model as a unique and a indexed column.
Bootstrap and other dependencies for the project were added as well.
During the process there were some problem where bootstrap module was
not available and thus webpacker compilation failed.

Compilation output was configured in 'webpacker.yml' to display more
verbose output and thus resolved with yarn adding the bootstrap and
other dependencies. The setting was reset as before.