ActionCable setup and form submit not working
Client side of the ActionCable setup has been mostly completed. Though
passing values from one side to the other is not configured correctly
as of now.
All current input submission attempts do not work. Javascript and
appropriate class and id naming conventions have been analyzed and
problem as of yet have not been indentified.
Jquery as required have been configured to be used in the webpacker
config file. Problems with the form was considered jquery's fault which
has worked but no effect on form could be seen.

Changes to the frontend calculation sees the use of helper methods in
future commits.