Testing and commenting out responsive chat  and small typo fixes
This commit fixes some of the typos in chat controller and adds
new jquery code to listen for keypress on enter. Currently the said
implementation is not working. Thus new jquery code is used cutting down
on current approach.

Responsive bootstrap rows and col-md-* were tested but current css
styles conflict with the design. The bootstrap code is currently
commented out and should be uncommented if responsive layout is
necessary.