Invitable functionality not implemented and message form
Devise Invitable is added to the application and mailer is configured
accordingly. Though invitation functionality is not available in the
application itself as of yet. This has been setup for adding the
functionality when the application works.

Message was intercepted incorrectly leading to the form not sending the
data to the resource. A submit button was added and js changes would be
made to app.