Message appear in the chat and update after refresh
The message intercept was not happening since the js function was not
being called. Current the message through the submit button is working
but the messages are not being updated realtime. The subscription to the
messages channel seems to be missing an argument and thus causing an
error.
The channel and subscription are to be checked next.