Messaging functionality with enter to send working correctly
The message box is now a single line text area that can submit input
with an enter from the user. The sent data doesn't update the messages
view as of now. Next problem to address should be to set the correct
number of arguments in the ActionCable broadcast so everything runs
correctly.